test: 3단계-시간관리 테스트 추가 및 단계별 테스트 클래스 분리
